+++ components/hero-section.tsx
@@ -1,0 +1,5 @@
+export function HeroSection({children}: {children: React.ReactNode}) {
+  return (
+    <section className="py-6 w-full flex flex-col gap-8">{children}</section>
+  )
+}
