+++ components/news-bit.tsx
@@ -1,7 +1,7 @@
-import {NewsList} from '@/components/news-list'
+import {NewsCard} from '@/components/news-card'
 import {CollectionSlug, DataFromCollectionSlug, PaginatedDocs} from 'payload'
 
-export function NewsSection({
+export function NewsBit({
   title,
   news,
   href
@@ -14,7 +14,7 @@ export function NewsSection({
     <div className="flex flex-col gap-4">
       <h2 className="text-2xl font-semibold">{title}</h2>
       <hr className="border-t border-gray-300" />
-      <NewsList news={news} href={href} />
+      <NewsCard news={news} href={href} />
     </div>
   )
 }
