--- components/section.tsx
@@ -1,7 +1,0 @@
-export function Section({children}: {children: React.ReactNode}) {
-  return (
-    <section className="py-6 w-full flex flex-col gap-8">
-      {children}
-    </section>
-  )
-}
